Fix: change decimal attributes to double for more precision

+++ powerplant-coding-challenge/Model/Powerplant.cs
@@ -10,7 +10,7 @@
     {
         public string Name { get; set; }
         public PowerplantType Type { get; set; }
-        public decimal Efficiency { get; set; }
+        public double Efficiency { get; set; }
         public int Pmin { get; set; }
         public int Pmax { get; set; }
